Rename the Google project to be consistent with others
