+++ PushSharp.Google/GcmServiceConnection.cs
@@ -9,7 +9,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.Serialization;
 
-namespace PushSharp.Google.Gcm
+namespace PushSharp.Google
 {
     public class GcmServiceConnectionFactory : IServiceConnectionFactory<GcmNotification>
     {
