+++ PushSharp.Google/Properties/AssemblyInfo.cs
