+++ PushSharp.Google/PushSharp.Google.csproj
@@ -5,7 +5,7 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{94F16497-471F-433F-A99E-C455FB2D7724}</ProjectGuid>
     <OutputType>Library</OutputType>
-    <RootNamespace>PushSharp.Google.Gcm</RootNamespace>
+    <RootNamespace>PushSharp.Google</RootNamespace>
     <AssemblyName>PushSharp.Google.Gcm</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
   </PropertyGroup>
