+++ PushSharp.Tests/PushSharp.Tests.csproj
@@ -56,13 +56,13 @@
       <Project>{A9D99F80-FEEB-4E74-96C5-66F17103C773}</Project>
       <Name>PushSharp.Apple</Name>
     </ProjectReference>
-    <ProjectReference Include="..\PushSharp.Google.Gcm\PushSharp.Google.Gcm.csproj">
-      <Project>{94F16497-471F-433F-A99E-C455FB2D7724}</Project>
-      <Name>PushSharp.Google.Gcm</Name>
-    </ProjectReference>
     <ProjectReference Include="..\PushSharp.Amazon\PushSharp.Amazon.csproj">
       <Project>{2468C63B-C964-4FC3-9B16-13DC17CF7D11}</Project>
       <Name>PushSharp.Amazon</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\PushSharp.Google\PushSharp.Google.csproj">
+      <Project>{94F16497-471F-433F-A99E-C455FB2D7724}</Project>
+      <Name>PushSharp.Google</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
